NN-2351-fix-enddate-bug
