+++ src/BulkAppointments/RecurringAppointments.js
@@ -10,14 +10,6 @@ const repeatsToMomentMap = {
   FORTNIGHTLY: 'weeks',
 }
 
-const repeatTypes = [
-  { value: 'WEEKLY', text: 'Weekly' },
-  { value: 'DAILY', text: 'Daily' },
-  { value: 'WEEKDAYS', text: 'Weekday (Monday to Friday)' },
-  { value: 'MONTHLY', text: 'Monthly' },
-  { value: 'FORTNIGHTLY', text: 'Fortnightly' },
-]
-
 const calculateEndDate = ({ startTime, repeats, numberOfTimes }) => {
   if (!startTime) return null
 
@@ -44,5 +36,4 @@ const recurringEndDate = values => {
 module.exports = {
   calculateEndDate,
   recurringEndDate,
-  repeatTypes,
 }
